# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Solid Tech Backend API",
    "description": "Backend API contract for frontend mock replacement.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/stock": {
      "get": {
        "tags": [
          "Stock"
        ],
        "summary": "Get Stock",
        "operationId": "get_stock_api_v1_stock_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Stock"
        ],
        "summary": "Create Drink",
        "operationId": "create_drink_api_v1_stock_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDrinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/stock/{drink_id}/quantity": {
      "patch": {
        "tags": [
          "Stock"
        ],
        "summary": "Update Stock Quantity",
        "operationId": "update_stock_quantity_api_v1_stock__drink_id__quantity_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "drink_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Drink Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDrinkQuantityRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stock/{drink_id}/alert-level": {
      "patch": {
        "tags": [
          "Stock"
        ],
        "summary": "Update Stock Alert Level",
        "operationId": "update_stock_alert_level_api_v1_stock__drink_id__alert_level_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "drink_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Drink Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDrinkAlertLevelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateDrinkRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "price": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Price"
          },
          "stock": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Stock"
          },
          "alertLevel": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Alertlevel"
          },
          "staffName": {
            "type": "string",
            "title": "Staffname"
          }
        },
        "type": "object",
        "required": [
          "name",
          "category",
          "price",
          "stock",
          "alertLevel",
          "staffName"
        ],
        "title": "CreateDrinkRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateDrinkAlertLevelRequest": {
        "properties": {
          "alertLevel": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Alertlevel"
          },
          "staffName": {
            "type": "string",
            "title": "Staffname"
          }
        },
        "type": "object",
        "required": [
          "alertLevel",
          "staffName"
        ],
        "title": "UpdateDrinkAlertLevelRequest"
      },
      "UpdateDrinkQuantityRequest": {
        "properties": {
          "newQuantity": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Newquantity"
          },
          "staffName": {
            "type": "string",
            "title": "Staffname"
          }
        },
        "type": "object",
        "required": [
          "newQuantity",
          "staffName"
        ],
        "title": "UpdateDrinkQuantityRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}